feature : createOrder의 비즈니스 로직 검증 추가

+++ src/main/java/com/limitedgoods/limitedgoods/order/entity/OrderItem.java
@@ -25,5 +25,5 @@ public class OrderItem {
 
     private int price;
 
-    private int lineTotalPrice;
+    private long lineTotalPrice;
 }
